Added the ArgumentHandler.py and adjusted the imagefs.py script so that a user now can use the script as follows:
python imagefs.py -s source -d destination/mountpoint -c conf.xml

This is more natural and offers the possibility to add more arguments to the script. Afterwards, while processing, sys.argv is reset in such a way that argv[1] contains the mountpoint, as prescribed by Fuse. The rest of the arguments are processed by the ArgumentHandler class and can be approached through getters.